Room implementation

+++ weatherlib/src/main/java/vboo/com/weatherlib/domain/model/City.kt
@@ -3,4 +3,4 @@ package vboo.com.weatherlib.domain.model
 /**
  * This class handles main city informations
  */
-data class City(val id: Int, val name: String, val country: String, val lat: Int, val lon: Int)
+data class City(val id: Int, val name: String, val country: String, val lat: Double, val lon: Double)
